Patch the Google App Engine memcache API into the module dictionary as 'memcache' so the Django memcache cache backend will work without modification.
This is possible because the Google App Engine memcache API is compatible with
the standard memcached API.

To use the memcache backend in your application simply add:
 CACHE_BACKEND = 'memcached://'
to your settings.py file. No IP address or port are required.

This code is based on a patch provided by Jonca Rafal and addresses Issue #57.


git-svn-id: http://google-app-engine-django.googlecode.com/svn/trunk@39 d4307497-c249-0410-99bd-594fbd7e173e